# Slightly improve `wafris-rb` response time.