--- a/.tool-versions
+++ b/.tool-versions
@@ -1,1 +1,1 @@
-ruby 3.2.2
+ruby 3.3.0-preview1