--- a/.tool-versions
+++ b/.tool-versions
@@ -1,1 +1,1 @@
-ruby 3.3.0-preview1
+ruby 3.1.3