--- a/lib/wafris/railtie.rb
+++ b/lib/wafris/railtie.rb
@@ -3,7 +3,11 @@
 module Wafris
   class Railtie < ::Rails::Railtie
     initializer 'wafris.middleware' do |app|
-      app.middleware.use(Wafris::Middleware)
+      if Rails.version > "6" && defined?(ActionDispatch::HostAuthorization)
+        app.middleware.insert_after ActionDispatch::HostAuthorization, Middleware
+      else
+        app.middleware.use(Wafris::Middleware)
+      end
     end
   end
 end