--- a/wafris.gemspec
+++ b/wafris.gemspec
@@ -23,6 +23,7 @@ Gem::Specification.new do |s|
   s.required_ruby_version = '>= 2.5'
 
   s.add_runtime_dependency 'connection_pool', '>= 2.3'
+  s.add_runtime_dependency 'hiredis-client', '>= 0.19'
   s.add_runtime_dependency 'rack', '>= 2.0'
   s.add_runtime_dependency 'redis', '>= 4.8.0'