The tests were performed using [Siege](https://github.com/JoeDog/siege), e.g:

**main branch:**
```
Transactions:                  12500 hits
Availability:                 100.00 %
Elapsed time:                  88.44 secs
Data transferred:             143.84 MB
Response time:                  0.18 secs
Transaction rate:             141.34 trans/sec
Throughput:                     1.63 MB/sec
Concurrency:                   24.96
Successful transactions:       12500
Failed transactions:               0
Longest transaction:            1.04
Shortest transaction:           0.01
```

**this branch:**
```
Transactions:                  12500 hits
Availability:                 100.00 %
Elapsed time:                  81.33 secs
Data transferred:             143.84 MB
Response time:                  0.16 secs 
Transaction rate:             153.69 trans/sec
Throughput:                     1.77 MB/sec
Concurrency:                   24.97
Successful transactions:       12500
Failed transactions:               0
Longest transaction:            0.37
Shortest transaction:           0.02
```

Some points to consider:

- [ ] update the docs on how to use `hiredis`
- [ ] remove the log suppressor
- [ ] add any missing specs