@ruyrocha 2023-12-29
@rmcastil, here is the PR with some changes for the gem.

I think I may be missing something due to the difference on numbers for the object allocations, and the overall speed was slightly improved.

I checked the local Redis server and didn't find anything obvious that would be impacting on performance, and I didn't look at the Lua part.