@ruyrocha 2024-01-29
hey @rmcastil, what are your thoughts on this? Perhaps I can break it down into smaller PRs.