@ruyrocha 2024-02-13
@rmcastil :+1: here for the other small PRs.

I'm pretty sure timeouts and connection pooling are safe, but is the core (Lua part) up to date so I can double-check the Redis commands?