@rmcastil 2024-02-13
Lua core is up to date but as long as eval works it should be good.
What’s the big benefit of the library?

On February 12, 2024, GitHub ***@***.***> wrote:
> @rmcastil <https://github.com/rmcastil> 👍 here for the other small
> PRs.
>
> I'm pretty sure timeouts and connection pooling are safe, but is the
> core (Lua part) up to date so I can double-check the Redis commands?
>

—
Reply to this email directly, view it on GitHub
<https://github.com/Wafris/wafris-rb/pull/69#issuecomment-1940251457>,
or unsubscribe <https://github.com/notifications/unsubscribe-
auth/AABF6M3YWHWYIVQ7XWMTQMTYTLD6NAVCNFSM6AAAAABBG2XRJOVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMYTSNBQGI2TCNBVG4>.
You are receiving this because you were mentioned.Message ID:
***@***.***>