@ruyrocha 2024-02-13
@rmcastil the main reasoning for speed is around the implementation (C vs Ruby).

There's this [public benchmark](https://github.com/redis/hiredis-rb?tab=readme-ov-file#benchmarks) from a while ago, and this [other one](https://gist.github.com/jgaskins/ecc0fc90d78cdec63c31e0ce5544faa1#file-output-m1-txt) with recent hardware but not so conclusive; I'm only stating they're not conclusive based on Redis commands for the implementation using `TCPSocket`.