@ruyrocha 2024-02-13
@rmcastil I keep asking myself what would be the best strategy to check the whole outcome, and I think Siege can simplify it; do you have any other suggestions to test the changes? They involve loading the middleware early on Rails, and the usage of `hiredis`.