@rmcastil 2024-03-25
Closing this as it's not critical to the future of core. Plus we've been discussing a different architecture approach internally.